Craft `spawn` with svm-cli
